requirements.txt
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.4.0
django-filter==24.3
djangorestframework-simplejwt


